pyproject.toml
[project]
name = "dsal"
version = "0.1.0"
description = "This is the corresponding .toml file required to create the project's uv environment."
requires-python = "==3.10"
dependencies = [
    'tensorflow[and-cuda]==2.15.1',
    'tensorflow-probability==0.23.0',
    'statsmodels',
    'dtaidistance',
    'plotly',
    'python-dotenv',
    'scikit-learn',
    'tqdm',
    'openpyxl',
]
